chore(cli): Move EAS remote build cache to a separated file
